+++ frontend/src/app/components/Button.module.css
@@ -5,4 +5,5 @@
   font-style: normal;
   font-weight: 500;
   line-height: normal;
+  cursor: pointer;
 }
