+++ frontend/src/app/components/Button.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import Link from "next/link";
 import React from "react";
 
 import styles from "./Button.module.css";
