+++ frontend/src/app/components/CustomDropdown.module.css
@@ -15,6 +15,7 @@
   display: flex;
   align-items: center;
   justify-content: space-between;
+  cursor: pointer;
 }
 
 .customDropdown {
