+++ frontend/src/app/components/FileUpload.tsx
@@ -2,10 +2,10 @@ import { ref, uploadBytes } from "firebase/storage";
 import Image from "next/image";
 import React, { ChangeEvent, useState } from "react";
 
+import { storage } from "../../../firebase/firebase";
 import { CreateFileObjectRequest, createFileObject } from "../api/fileApi";
 
 import styles from "./FileUpload.module.css";
-import { storage } from "../../../firebase/firebase";
 
 type FileUploadProps = {
   onClose: () => void;
