+++ frontend/src/app/components/OnboardingOption.module.css
@@ -5,6 +5,7 @@
   padding: 12px 16px;
   justify-content: center;
   align-items: center;
+  cursor: pointer;
 
   border-radius: 8px;
   border: 1.5px solid var(--Stroke-2, #d0d7de);
