+++ frontend/src/app/components/SignUpForm/page.module.css
@@ -1,5 +1,5 @@
 .page {
-  background-image: url("../../../public/signUpBackground.svg");
+  background-image: url("../../../../public/signUpBackground.svg");
   background-size: cover;
   min-height: 100vh;
   width: 100%;
@@ -215,7 +215,7 @@
 }
 
 .error {
-  color: #60696f;
+  color: #b80037;
   font-family: "Albert Sans";
   font-size: 14px;
   margin-bottom: 4px;
